# vuelo/swagger.yaml
servers:
- {url: 'http://localhost:8080/api/'}
openapi: 3.0.0
paths:
  /tripulante/:
    get:
      summary: /tripulante getAll
      operationId: /tripulante_getAll
      responses:
        '200': {description: Response 200}
      tags: [/tripulante]
  /vuelo/{key}:
    get:
      summary: /vuelo getByKey
      operationId: /vuelo_getByKey
      responses:
        '200': {description: Response 200}
      parameters:
      - schema: {type: string}
        in: path
        name: key
        required: true
      tags: [/vuelo]
    delete:
      summary: /vuelo delete
      operationId: /vuelo_delete
      responses:
        '200': {description: Response 200}
      parameters:
      - schema: {type: string}
        in: path
        name: key
        required: true
      tags: [/vuelo]
    put:
      summary: /vuelo edit
      requestBody:
        content:
          application/json:
            schema: {type: object}
      operationId: /vuelo_edit
      responses:
        '200': {description: Response 200}
      parameters:
      - schema: {type: string}
        in: path
        name: key
        required: true
      tags: [/vuelo]
  /vuelo/registro:
    post:
      summary: /vuelo register
      requestBody:
        content:
          application/json:
            schema: {type: object}
      operationId: /vuelo_register
      responses:
        '200': {description: Response 200}
      tags: [/vuelo]
  /vuelo/:
    get:
      summary: /vuelo getAll
      operationId: /vuelo_getAll
      responses:
        '200': {description: Response 200}
      tags: [/vuelo]
  /tripulante/{key}:
    get:
      summary: /tripulante getByKey
      operationId: /tripulante_getByKey
      responses:
        '200': {description: Response 200}
      parameters:
      - schema: {type: string}
        in: path
        name: key
        required: true
      tags: [/tripulante]
info: {description: FuourTeam API description, title: FourTeam API, version: 1.0.1}
tags:
- {name: /tripulante, description: /tripulante_descripcion}
- {name: /vuelo, description: /vuelo_descripcion}